scripts/requirements.txt
pyspark
psycopg2-binary
spotipy
tabulate
dotenv
celery
findspark